refactor: remove unnecessary hovering toolbar controls component
