+++ packages/plugins/text/src/components/hovering-toolbar.tsx
@@ -1,19 +1,82 @@
+import * as R from 'ramda'
 import React from 'react'
 import { Editor as SlateEditor, Range } from 'slate'
 import { ReactEditor, useSlate } from 'slate-react'
 
-import type { HoveringToolbarProps } from '../types'
+import { useToolbarControls } from '../hooks/use-toolbar-controls'
+import type {
+  HoveringToolbarProps,
+  NestedControlButton,
+  TextEditorControl,
+} from '../types'
 import { isTouchDevice } from '../utils/is-touch-device'
-import { HoveringToolbarControls } from './hovering-toolbar-controls'
+import { HoveringToolbarButton } from './hovering-toolbar-button'
 import { InlineOverlay, InlineOverlayPosition } from './inline-overlay'
 
 const initialPosition = isTouchDevice()
   ? InlineOverlayPosition.below
   : InlineOverlayPosition.above
 
+function isNestedControlButton(
+  control: TextEditorControl
+): control is NestedControlButton {
+  return R.has('children', control)
+}
+
 export function HoveringToolbar({ config }: HoveringToolbarProps) {
+  const [subMenu, setSubMenu] = React.useState<number>()
   const editor = useSlate()
   const { selection } = editor
+  const controls = useToolbarControls(config)
+
+  if (typeof subMenu !== 'number') {
+    return (
+      <InlineOverlay
+        initialPosition={initialPosition}
+        hidden={
+          !selection ||
+          !ReactEditor.isFocused(editor) ||
+          Range.isCollapsed(selection) ||
+          SlateEditor.string(editor, selection) === ''
+        }
+      >
+        {controls.map((control, index) => (
+          <HoveringToolbarButton
+            active={control.isActive(editor)}
+            theme={config.theme}
+            title={control.title}
+            onMouseDown={(event) => {
+              event.preventDefault()
+              isNestedControlButton(control)
+                ? setSubMenu(index)
+                : control.onClick(editor)
+            }}
+            key={index}
+          >
+            {control.renderIcon()}
+          </HoveringToolbarButton>
+        ))}
+      </InlineOverlay>
+    )
+  }
+
+  const activeControl = controls[subMenu]
+
+  if (!isNestedControlButton(activeControl)) return null
+
+  const closeSubMenuControl = {
+    isActive() {
+      return false
+    },
+    renderIcon() {
+      return activeControl.renderCloseMenuIcon()
+    },
+    onClick() {
+      setSubMenu(undefined)
+    },
+    title: activeControl.closeMenuTitle,
+  }
+  const subMenuControls = [...activeControl.children, closeSubMenuControl]
 
   return (
     <InlineOverlay
@@ -25,7 +88,20 @@ export function HoveringToolbar({ config }: HoveringToolbarProps) {
         SlateEditor.string(editor, selection) === ''
       }
     >
-      <HoveringToolbarControls editor={editor} config={config} />
+      {subMenuControls.map((control, index) => (
+        <HoveringToolbarButton
+          active={control.isActive(editor)}
+          theme={config.theme}
+          title={control.title}
+          onMouseDown={(event) => {
+            event.preventDefault()
+            control.onClick(editor)
+          }}
+          key={index}
+        >
+          {control.renderIcon()}
+        </HoveringToolbarButton>
+      ))}
     </InlineOverlay>
   )
 }
