+++ packages/plugins/text/src/types/hovering-toolbar.ts
@@ -1,26 +1,11 @@
 import { ListsEditor } from '@prezly/slate-lists'
-import { MouseEventHandler } from 'react'
-import { Editor as SlateEditor, BaseEditor } from 'slate'
+import { BaseEditor } from 'slate'
 import { ReactEditor } from 'slate-react'
 
 import { TextPluginConfig } from '.'
 
 export interface HoveringToolbarProps {
   config: TextPluginConfig
-}
-
-export interface HoveringToolbarControlsProps {
-  editor: SlateEditor
-  config: TextPluginConfig
-}
-
-export interface HoveringToolbarButtonProps {
-  editor: SlateEditor
-  config: TextPluginConfig
-  index?: number
-  control: TextEditorControl
-  setSubMenu: React.Dispatch<React.SetStateAction<number | undefined>>
-  onMouseDown: MouseEventHandler
 }
 
 export type TextEditorControl = ControlButton | NestedControlButton
